pyproject.toml
[tool.poetry]
name = "chirpstack_client"
version = "0.0.1"
description = ""
authors = ["Shivam Shandilya <vyogami@gmail.com>"]
readme = "README.md"
packages = [
  {include = "chirpstack_client", from = "src"}
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
chirpstack-api = "^4.9.0"
grpcio = "^1.66.2"
protobuf = "^5.28.2"
cryptography = "^43.0.1"
pycryptodome = "^3.21.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
rich = "^13.8.0"
python-dotenv = "^1.0.1"
grpcio-tools = "^1.66.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py37"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    "TRY",
    "A003",
    "B904",
    "RUF010",
]

[tool.ruff.format]
preview = true


[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
"src/chirpstack_client/proto/*"= ["S101"]
